Move models to namespace
